// simulation/types.ts
export type GovernorType = 'oz' | 'bravo'

interface SimulationConfigBase {
  type: 'executed' | 'proposed' | 'new'
  daoName: string // e.g. 'Compound' or 'Uniswap'
  governorAddress: string // address of the governor
  governorType: GovernorType
}

export interface SimulationConfigExecuted extends SimulationConfigBase {
  type: 'executed'
  proposalId: BigNumberish // ID of the executed proposal
}

export interface SimulationConfigProposed extends SimulationConfigBase {
  type: 'proposed'
  proposalId: BigNumberish // ID of the executed proposal
}

export interface SimulationConfigNew extends SimulationConfigBase {
  type: 'new'
  targets: string[]
  values: BigNumberish[]
  signatures: string[]
  calldatas: string[]
  description: string
}

export type SimulationConfig = SimulationConfigExecuted | SimulationConfigProposed | SimulationConfigNew

export interface SimulationResult {
  sim: TenderlySimulation
  proposal: ProposalEvent
  latestBlock: Block
}
// ...
export interface ProposalEvent {
  id?: BigNumber // Bravo governor
  proposalId?: BigNumber // OZ governor
  proposer: string
  startBlock: BigNumber
  endBlock: BigNumber
  description: string
  targets: string[]
  values: BigNumber[]
  signatures: string[]
  calldatas: string[]
}
// ...
export type StorageEncodingResponse = {
  stateOverrides: {
    // these keys are the contract addresses, all lower case
    [key: string]: {
      value: {
        // these are the slot numbers, as 32 byte hex strings
        [key: string]: string
      }
    }
  }
}

type StateObject = {
  balance?: string
  code?: string
  storage?: Record<string, string>
}

type ContractObject = {
  contractName: string
  source: string
  sourcePath: string
  compiler: {
    name: 'solc'
    version: string
  }
  networks: Record<
    string,
    {
      events?: Record<string, string>
      links?: Record<string, string>
      address: string
      transactionHash?: string
    }
  >
}

export type TenderlyPayload = {
  network_id: '1' | '3' | '4' | '5' | '42'
  block_number?: number
  transaction_index?: number
  from: string
  to: string
  input: string
  gas: number
  gas_price?: string
  value?: string
  simulation_type?: 'full' | 'quick'
  save?: boolean
  save_if_fails?: boolean
  state_objects?: Record<string, StateObject>
  contracts?: ContractObject[]
  block_header?: {
    number?: string
    timestamp?: string
  }
  generate_access_list?: boolean
}

// --- Tenderly types, Response ---
// NOTE: These type definitions were autogenerated using https://app.quicktype.io/, so are almost
// certainly not entirely accurate (and they have some interesting type names)

export interface TenderlySimulation {
  transaction: Transaction
  simulation: Simulation
  contracts: TenderlyContract[]
  generated_access_list: GeneratedAccessList[]
}

export interface TenderlyContract {
  id: string
  contract_id: string
  balance: string
  network_id: string
  public: boolean
  boolean
  verified_by: string
  verification_date: null
  address: string
  contract_name: string
  ens_domain: null
  type: string
  evm_version: string
  compiler_version: string
  optimizations_used: boolean
  optimization_runs: number
  libraries: null
  data: Data
  creation_block: number
  creation_tx: string
  creator_address: string
  created_at: Date
  number_of_watches: null
  language: string
  in_project: boolean
  number_of_files: number
  standard?: string
  standards?: string[]
  token_data?: TokenData
}

interface Data {
  main_contract: number
  contract_info: ContractInfo[]
  abi: ABI[]
  raw_abi: null
}

interface ABI {
  type: ABIType
  name: string
  constant: boolean
  anonymous: boolean
  inputs: SoltypeElement[]
  outputs: Output[] | null
}
// ...
interface ContractInfo {
  id: number
  path: string
  name: string
  source: string
}

interface TokenData {
  symbol: string
  name: string
  decimals: number
}

interface GeneratedAccessList {
  address: string
  storage_keys: string[]
}

interface Simulation {
  id: string
  project_id: string
  owner_id: string
  network_id: string
  block_number: number
  transaction_index: number
  from: string
  to: string
  input: string
  gas: number
  gas_price: string
  value: string
  method: string
  status: boolean
  access_list: null
  queue_origin: string
  created_at: Date
}

interface Transaction {
  hash: From
  block_hash: string
  block_number: number
  from: From
  gas: number
  gas_price: number
  gas_fee_cap: number
  gas_tip_cap: number
  cumulative_gas_used: number
  gas_used: number
  effective_gas_price: number
  input: string
  nonce: number
  to: From
  index: number
  value: string
  access_list: null
  status: boolean
  addresses: string[]
  contract_ids: string[]
  network_id: string
  function_selector: string
  transaction_info: TransactionInfo
  timestamp: Date
  method: string
  decoded_input: null
}

interface TransactionInfo {
  contract_id: string
  block_number: number
  transaction_id: From
  contract_address: From
  method: string
  parameters: null
  intrinsic_gas: number
  refund_gas: number
  call_trace: CallTrace
  stack_trace: null | StackTrace[]
  logs: Log[] | null
  state_diff: StateDiff[]
  raw_state_diff: null
  console_logs: null
  created_at: Date
}